+++ CHANGELOG.md
@@ -264,3 +264,7 @@
 # 0.0.1 (Update 19/05/2023)
 - Modul Kehadiran Bulanan, Office, Produksi
     - fix issue tambah absensi manual dan hapus absensi
+
+# 0.0.1 (Update 22/05/2023)
+- Modul Pengeluaran
+    - pengeluaran untuk toko bulanan & office